Newly designed website
It was about time we move forward with our project website content
and layout. The new page is based on the lanyon theme for jekyll
and run on the github pages infrastructure for easier hosting.

The content of the old page have been moved over and extended or
trimmed as seen fit. There is still some work to do but it should
be on the same level or better compared to our old website.